Resolve the root of one project apart from the paths built from it
An application written as several projects has one source root per
project and a single directory to write all of their paths relative to,
worked out from the roots together. Saying a project's root as a
directory rather than only as the paths made from it is what lets that
be asked at all.

The comparing of directories moves out with it. The same three
questions - normalize, share a prefix, is this a file system root - are
asked of one project's directories and of the roots of all of them, so
they belong beside neither.
